fix: fixed the menu structure and added a checkout entry
