+++ apps/web/components/settings/checkout/lang.json
@@ -1,0 +1,8 @@
+{
+  "en": {
+    "checkout": "Checkout"
+  },
+  "de": {
+    "checkout": "Checkout"
+  }
+}
