+++ apps/web/components/settings/checkout/other-settings/View.vue
@@ -1,0 +1,7 @@
+<template>
+  <SiteConfigurationView>
+    <template #setting-title> Other Settings </template>
+  </SiteConfigurationView>
+</template>
+
+<script setup lang="ts"></script>
