+++ apps/web/components/settings/checkout/other-settings/lang.json
@@ -1,0 +1,8 @@
+{
+  "en": {
+    "other-settings": "Other Settings"
+  },
+  "de": {
+    "other-settings": "Other Settings"
+  }
+}
