--- apps/web/components/settings/general/vat-validation/lang.json
@@ -1,8 +1,0 @@
-{
-  "en": {
-    "vat-validation": "Vat Validation"
-  },
-  "de": {
-    "vat-validation": "Vat Validation"
-  }
-}
